REVSDL-2379 Mobile command factory for VR Plugin
rename ActiveServiceRequest to ActivateServiceRequest
