+++ src/components/vr_cooperation/CMakeLists.txt
@@ -41,7 +41,7 @@ set (SOURCES
     ./src/base_command_request.cc
     ./src/register_service_request.cc
     ./src/unregister_service_request.cc
-    ./src/active_service_request.cc
+    ./src/activate_service_request.cc
 )
 
 set (LIBRARIES
